@@ -1,9 +1,15 @@
-// swift-tools-version:5.10
+// swift-tools-version:6.0
 import PackageDescription
 
+#if compiler(>=6.1)
+let swiftSettings: [SwiftSetting] = []
+#else
 let swiftSettings: [SwiftSetting] = [
-    .enableUpcomingFeature("StrictConcurrency"),
+    // Sadly the 6.0 compiler concurrency checker finds false positives.
+    // To be able to compile, lets reduce the language version down to 5 for 6.0 only.
+    .swiftLanguageMode(.v5)
 ]
+#endif
 
 let package = Package(
     name: "postgres-nio",
