@@ -2,12 +2,13 @@
 import _ConnectionPoolTestUtils
 import Atomics
 import NIOEmbedded
-import XCTest
+import Testing
 
-@available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
-final class ConnectionPoolTests: XCTestCase {
 
-    func test1000ConsecutiveRequestsOnSingleConnection() async {
+@Suite struct ConnectionPoolTests {
+
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func test1000ConsecutiveRequestsOnSingleConnection() async {
         let factory = MockConnectionFactory<ContinuousClock>()
 
         var config = ConnectionPoolConfiguration()
@@ -34,35 +35,35 @@ final class ConnectionPoolTests: XCTestCase {
             let createdConnection = await factory.nextConnectAttempt { _ in
                 return 1
             }
-            XCTAssertNotNil(createdConnection)
 
             do {
                 for _ in 0..<1000 {
-                    async let connectionFuture = try await pool.leaseConnection()
+                    async let connectionFuture = pool.leaseConnection()
                     var connectionLease: ConnectionLease<MockConnection>?
-                    XCTAssertEqual(factory.pendingConnectionAttemptsCount, 0)
+                    #expect(factory.pendingConnectionAttemptsCount == 0)
                     connectionLease = try await connectionFuture
-                    XCTAssertNotNil(connectionLease)
-                    XCTAssert(createdConnection === connectionLease?.connection)
+                    #expect(connectionLease != nil)
+                    #expect(createdConnection === connectionLease?.connection)
 
                     connectionLease?.release()
                 }
             } catch {
-                XCTFail("Unexpected error: \(error)")
+                Issue.record("Unexpected error: \(error)")
             }
 
             taskGroup.cancelAll()
 
-            XCTAssertEqual(factory.pendingConnectionAttemptsCount, 0)
+            #expect(factory.pendingConnectionAttemptsCount == 0)
             for connection in factory.runningConnections {
                 connection.closeIfClosing()
             }
         }
 
-        XCTAssertEqual(factory.runningConnections.count, 0)
+        #expect(factory.runningConnections.count == 0)
     }
 
-    func testShutdownPoolWhileConnectionIsBeingCreated() async {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testShutdownPoolWhileConnectionIsBeingCreated() async {
         let clock = MockClock()
         let factory = MockConnectionFactory<MockClock>()
 
@@ -107,7 +108,8 @@ final class ConnectionPoolTests: XCTestCase {
         struct ConnectionCreationError: Error {}
     }
 
-    func testShutdownPoolWhileConnectionIsBackingOff() async {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testShutdownPoolWhileConnectionIsBackingOff() async {
         let clock = MockClock()
         let factory = MockConnectionFactory<MockClock>()
 
@@ -142,7 +144,8 @@ final class ConnectionPoolTests: XCTestCase {
         struct ConnectionCreationError: Error {}
     }
 
-    func testConnectionHardLimitIsRespected() async {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testConnectionHardLimitIsRespected() async {
         let factory = MockConnectionFactory<ContinuousClock>()
 
         var mutableConfig = ConnectionPoolConfiguration()
@@ -171,21 +174,21 @@ final class ConnectionPoolTests: XCTestCase {
         await withTaskGroup(of: Void.self) { taskGroup in
             taskGroup.addTask_ {
                 await pool.run()
-                XCTAssertFalse(hasFinished.compareExchange(expected: false, desired: true, ordering: .relaxed).original)
+                #expect(hasFinished.compareExchange(expected: false, desired: true, ordering: .relaxed).original == false)
             }
 
             taskGroup.addTask_ {
                 var usedConnectionIDs = Set<Int>()
                 for _ in 0..<config.maximumConnectionHardLimit {
                     await factory.nextConnectAttempt { connectionID in
-                        XCTAssertTrue(usedConnectionIDs.insert(connectionID).inserted)
+                        #expect(usedConnectionIDs.insert(connectionID).inserted == true)
                         createdConnections.wrappingIncrement(ordering: .relaxed)
                         return 1
                     }
                 }
 
 
-                XCTAssertEqual(factory.pendingConnectionAttemptsCount, 0)
+                #expect(factory.pendingConnectionAttemptsCount == 0)
             }
 
             let (stream, continuation) = AsyncStream.makeStream(of: Void.self)
@@ -196,7 +199,7 @@ final class ConnectionPoolTests: XCTestCase {
                         let connectionLease = try await pool.leaseConnection()
                         connectionLease.release()
                     } catch {
-                        XCTFail("Unexpected error: \(error)")
+                        Issue.record("Unexpected error: \(error)")
                     }
                     continuation.yield()
                 }
@@ -209,18 +212,19 @@ final class ConnectionPoolTests: XCTestCase {
 
             taskGroup.cancelAll()
 
-            XCTAssertFalse(hasFinished.load(ordering: .relaxed))
+            #expect(hasFinished.load(ordering: .relaxed) == false)
             for connection in factory.runningConnections {
                 connection.closeIfClosing()
             }
         }
 
-        XCTAssertEqual(createdConnections.load(ordering: .relaxed), config.maximumConnectionHardLimit)
-        XCTAssert(hasFinished.load(ordering: .relaxed))
-        XCTAssertEqual(factory.runningConnections.count, 0)
+        #expect(createdConnections.load(ordering: .relaxed) == config.maximumConnectionHardLimit)
+        #expect(hasFinished.load(ordering: .relaxed) == true)
+        #expect(factory.runningConnections.count == 0)
     }
 
-    func testKeepAliveWorks() async throws {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testKeepAliveWorks() async throws {
         let clock = MockClock()
         let factory = MockConnectionFactory<MockClock>()
         let keepAliveDuration = Duration.seconds(30)
@@ -255,7 +259,7 @@ final class ConnectionPoolTests: XCTestCase {
             }
 
             let connectionLease = try await connectionLeaseFuture
-            XCTAssert(connection === connectionLease.connection)
+            #expect(connection === connectionLease.connection)
 
             connectionLease.release()
 
@@ -265,11 +269,11 @@ final class ConnectionPoolTests: XCTestCase {
             var expectedInstants: Set<MockClock.Instant> = [.init(keepAliveDuration), .init(config.idleTimeout)]
             let deadline1 = await clock.nextTimerScheduled()
             print(deadline1)
-            XCTAssertNotNil(expectedInstants.remove(deadline1))
+            #expect(expectedInstants.remove(deadline1) != nil)
             let deadline2 = await clock.nextTimerScheduled()
             print(deadline2)
-            XCTAssertNotNil(expectedInstants.remove(deadline2))
-            XCTAssert(expectedInstants.isEmpty)
+            #expect(expectedInstants.remove(deadline2) != nil)
+            #expect(expectedInstants.isEmpty == true)
 
             // move clock forward to keep alive
             let newTime = clock.now.advanced(by: keepAliveDuration)
@@ -278,14 +282,14 @@ final class ConnectionPoolTests: XCTestCase {
 
             await keepAlive.nextKeepAlive { keepAliveConnection in
                 defer { print("keep alive 1 has run") }
-                XCTAssertTrue(keepAliveConnection === connectionLease.connection)
+                #expect(keepAliveConnection === connectionLease.connection)
                 return true
             }
 
             // keep alive 2
 
             let deadline3 = await clock.nextTimerScheduled()
-            XCTAssertEqual(deadline3, clock.now.advanced(by: keepAliveDuration))
+            #expect(deadline3 == clock.now.advanced(by: keepAliveDuration))
             print(deadline3)
 
             // race keep alive vs timeout
@@ -299,7 +303,8 @@ final class ConnectionPoolTests: XCTestCase {
         }
     }
 
-    func testKeepAliveOnClose() async throws {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testKeepAliveOnClose() async throws {
         let clock = MockClock()
         let factory = MockConnectionFactory<MockClock>()
         let keepAliveDuration = Duration.seconds(20)
@@ -334,7 +339,7 @@ final class ConnectionPoolTests: XCTestCase {
             }
 
             let connectionLease = try await connectionLeaseFuture
-            XCTAssert(connection === connectionLease.connection)
+            #expect(connection === connectionLease.connection)
 
             connectionLease.release()
 
@@ -344,38 +349,38 @@ final class ConnectionPoolTests: XCTestCase {
             var expectedInstants: Set<MockClock.Instant> = [.init(keepAliveDuration), .init(config.idleTimeout)]
             let deadline1 = await clock.nextTimerScheduled()
             print(deadline1)
-            XCTAssertNotNil(expectedInstants.remove(deadline1))
+            #expect(expectedInstants.remove(deadline1) != nil)
             let deadline2 = await clock.nextTimerScheduled()
             print(deadline2)
-            XCTAssertNotNil(expectedInstants.remove(deadline2))
-            XCTAssert(expectedInstants.isEmpty)
+            #expect(expectedInstants.remove(deadline2) != nil)
+            #expect(expectedInstants.isEmpty)
 
             // move clock forward to keep alive
             let newTime = clock.now.advanced(by: keepAliveDuration)
             clock.advance(to: newTime)
 
             await keepAlive.nextKeepAlive { keepAliveConnection in
-                XCTAssertTrue(keepAliveConnection === connectionLease.connection)
+                #expect(keepAliveConnection === connectionLease.connection)
                 return true
             }
 
             // keep alive 2
             let deadline3 = await clock.nextTimerScheduled()
-            XCTAssertEqual(deadline3, clock.now.advanced(by: keepAliveDuration))
+            #expect(deadline3 == clock.now.advanced(by: keepAliveDuration))
             clock.advance(to: clock.now.advanced(by: keepAliveDuration))
 
             let failingKeepAliveDidRun = ManagedAtomic(false)
             // the following keep alive should not cause a crash
             _ = try? await keepAlive.nextKeepAlive { keepAliveConnection in
                 defer { 
-                    XCTAssertFalse(failingKeepAliveDidRun
-                        .compareExchange(expected: false, desired: true, ordering: .relaxed).original)
+                    #expect(failingKeepAliveDidRun
+                        .compareExchange(expected: false, desired: true, ordering: .relaxed).original == false)
                 }
-                XCTAssertTrue(keepAliveConnection === connectionLease.connection)
+                #expect(keepAliveConnection === connectionLease.connection)
                 keepAliveConnection.close()
                 throw CancellationError() // any error 
             } // will fail and it's expected
-            XCTAssertTrue(failingKeepAliveDidRun.load(ordering: .relaxed))
+            #expect(failingKeepAliveDidRun.load(ordering: .relaxed) == true)
 
             taskGroup.cancelAll()
 
@@ -385,7 +390,8 @@ final class ConnectionPoolTests: XCTestCase {
         }
     }
 
-    func testKeepAliveWorksRacesAgainstShutdown() async throws {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testKeepAliveWorksRacesAgainstShutdown() async throws {
         let clock = MockClock()
         let factory = MockConnectionFactory<MockClock>()
         let keepAliveDuration = Duration.seconds(30)
@@ -420,7 +426,7 @@ final class ConnectionPoolTests: XCTestCase {
             }
 
             let connectionLease = try await connectionLeaseFuture
-            XCTAssert(connection === connectionLease.connection)
+            #expect(connection === connectionLease.connection)
 
             connectionLease.release()
 
@@ -430,17 +436,17 @@ final class ConnectionPoolTests: XCTestCase {
             var expectedInstants: Set<MockClock.Instant> = [.init(keepAliveDuration), .init(config.idleTimeout)]
             let deadline1 = await clock.nextTimerScheduled()
             print(deadline1)
-            XCTAssertNotNil(expectedInstants.remove(deadline1))
+            #expect(expectedInstants.remove(deadline1) != nil)
             let deadline2 = await clock.nextTimerScheduled()
             print(deadline2)
-            XCTAssertNotNil(expectedInstants.remove(deadline2))
-            XCTAssert(expectedInstants.isEmpty)
+            #expect(expectedInstants.remove(deadline2) != nil)
+            #expect(expectedInstants.isEmpty)
 
             clock.advance(to: clock.now.advanced(by: keepAliveDuration))
 
             await keepAlive.nextKeepAlive { keepAliveConnection in
                 defer { print("keep alive 1 has run") }
-                XCTAssertTrue(keepAliveConnection === connectionLease.connection)
+                #expect(keepAliveConnection === connectionLease.connection)
                 return true
             }
 
@@ -453,7 +459,8 @@ final class ConnectionPoolTests: XCTestCase {
         }
     }
 
-    func testCancelConnectionRequestWorks() async throws {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testCancelConnectionRequestWorks() async throws {
         let clock = MockClock()
         let factory = MockConnectionFactory<MockClock>()
         let keepAliveDuration = Duration.seconds(30)
@@ -501,9 +508,9 @@ final class ConnectionPoolTests: XCTestCase {
             let taskResult = await leaseTask.result
             switch taskResult {
             case .success:
-                XCTFail("Expected task failure")
+                Issue.record("Expected task failure")
             case .failure(let failure):
-                XCTAssertEqual(failure as? ConnectionPoolError, .requestCancelled)
+                #expect(failure as? ConnectionPoolError == .requestCancelled)
             }
 
             taskGroup.cancelAll()
@@ -513,7 +520,8 @@ final class ConnectionPoolTests: XCTestCase {
         }
     }
 
-    func testLeasingMultipleConnectionsAtOnceWorks() async throws {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testLeasingMultipleConnectionsAtOnceWorks() async throws {
         let clock = MockClock()
         let factory = MockConnectionFactory<MockClock>()
         let keepAliveDuration = Duration.seconds(30)
@@ -562,7 +570,7 @@ final class ConnectionPoolTests: XCTestCase {
             }
 
             // Ensure that we got 4 distinct connections
-            XCTAssertEqual(Set(connectionLeases.lazy.map(\.connection.id)).count, 4)
+            #expect(Set(connectionLeases.lazy.map(\.connection.id)).count == 4)
 
             // release all 4 leased connections
             for lease in connectionLeases {
@@ -577,7 +585,8 @@ final class ConnectionPoolTests: XCTestCase {
         }
     }
 
-    func testLeasingConnectionAfterShutdownIsInvokedFails() async throws {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testLeasingConnectionAfterShutdownIsInvokedFails() async throws {
         let clock = MockClock()
         let factory = MockConnectionFactory<MockClock>()
         let keepAliveDuration = Duration.seconds(30)
@@ -618,10 +627,10 @@ final class ConnectionPoolTests: XCTestCase {
 
             do {
                 _ = try await pool.leaseConnection()
-                XCTFail("Expected a failure")
+                Issue.record("Expected a failure")
             } catch {
                 print("failed")
-                XCTAssertEqual(error as? ConnectionPoolError, .poolShutdown)
+                #expect(error as? ConnectionPoolError == .poolShutdown)
             }
 
             print("will close connections: \(factory.runningConnections)")
@@ -632,7 +641,8 @@ final class ConnectionPoolTests: XCTestCase {
         }
     }
 
-    func testLeasingConnectionsAfterShutdownIsInvokedFails() async throws {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testLeasingConnectionsAfterShutdownIsInvokedFails() async throws {
         let clock = MockClock()
         let factory = MockConnectionFactory<MockClock>()
         let keepAliveDuration = Duration.seconds(30)
@@ -680,9 +690,9 @@ final class ConnectionPoolTests: XCTestCase {
             for request in requests {
                 do {
                     _ = try await request.future.success
-                    XCTFail("Expected a failure")
+                    Issue.record("Expected a failure")
                 } catch {
-                    XCTAssertEqual(error as? ConnectionPoolError, .poolShutdown)
+                    #expect(error as? ConnectionPoolError == .poolShutdown)
                 }
             }
 
@@ -693,7 +703,8 @@ final class ConnectionPoolTests: XCTestCase {
         }
     }
 
-    func testLeasingMultipleStreamsFromOneConnectionWorks() async throws {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testLeasingMultipleStreamsFromOneConnectionWorks() async throws {
         let clock = MockClock()
         let factory = MockConnectionFactory<MockClock>()
         let keepAliveDuration = Duration.seconds(30)
@@ -737,7 +748,7 @@ final class ConnectionPoolTests: XCTestCase {
             }
 
             // Ensure that all requests got the same connection
-            XCTAssertEqual(Set(connectionLeases.lazy.map(\.connection.id)).count, 1)
+            #expect(Set(connectionLeases.lazy.map(\.connection.id)).count == 1)
 
             // release all 10 leased streams
             for lease in connectionLeases {
@@ -758,7 +769,8 @@ final class ConnectionPoolTests: XCTestCase {
         }
     }
 
-    func testIncreasingAvailableStreamsWorks() async throws {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testIncreasingAvailableStreamsWorks() async throws {
         let clock = MockClock()
         let factory = MockConnectionFactory<MockClock>()
         let keepAliveDuration = Duration.seconds(30)
@@ -808,7 +820,7 @@ final class ConnectionPoolTests: XCTestCase {
             }
 
             // Ensure that all requests got the same connection
-            XCTAssertEqual(Set(connectionLease.lazy.map(\.connection.id)).count, 1)
+            #expect(Set(connectionLease.lazy.map(\.connection.id)).count == 1)
 
             requests = (22..<42).map { ConnectionFuture(id: $0) }
             pool.leaseConnections(requests)
