@@ -1,29 +1,29 @@
 @testable import _ConnectionPoolModule
 import _ConnectionPoolTestUtils
-import XCTest
+import Testing
 
-final class ConnectionRequestTests: XCTestCase {
+@Suite struct ConnectionRequestTests {
 
-    func testHappyPath() async throws {
+    @Test func testHappyPath() async throws {
         let mockConnection = MockConnection(id: 1)
         let lease = try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<ConnectionLease<MockConnection>, any Error>) in
             let request = ConnectionRequest(id: 42, continuation: continuation)
-            XCTAssertEqual(request.id, 42)
+            #expect(request.id == 42)
             let lease = ConnectionLease(connection: mockConnection) { _ in }
             continuation.resume(with: .success(lease))
         }
 
-        XCTAssert(lease.connection === mockConnection)
+        #expect(lease.connection === mockConnection)
     }
 
-    func testSadPath() async throws {
+    @Test func testSadPath() async throws {
         do {
             _ = try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<MockConnection, any Error>) in
                 continuation.resume(with: .failure(ConnectionPoolError.requestCancelled))
             }
-            XCTFail("This point should not be reached")
+            Issue.record("This point should not be reached")
         } catch {
-            XCTAssertEqual(error as? ConnectionPoolError, .requestCancelled)
+            #expect(error as? ConnectionPoolError == .requestCancelled)
         }
     }
 }
