@@ -1,11 +1,12 @@
 import _ConnectionPoolModule
 import _ConnectionPoolTestUtils
-import XCTest
+import Testing
 
-@available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
-final class NoKeepAliveBehaviorTests: XCTestCase {
-    func testNoKeepAlive() {
+
+@Suite struct NoKeepAliveBehaviorTests {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testNoKeepAlive() {
         let keepAliveBehavior = NoOpKeepAliveBehavior(connectionType: MockConnection.self)
-        XCTAssertNil(keepAliveBehavior.keepAliveFrequency)
+        #expect(keepAliveBehavior.keepAliveFrequency == nil)
     }
 }
