@@ -1,29 +1,30 @@
 @testable import _ConnectionPoolModule
 import _ConnectionPoolTestUtils
-import XCTest
+import Testing
 
-@available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
-final class PoolStateMachine_RequestQueueTests: XCTestCase {
+@Suite struct PoolStateMachine_RequestQueueTests {
 
     typealias TestQueue = TestPoolStateMachine.RequestQueue
 
-    func testHappyPath() {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testHappyPath() {
         var queue = TestQueue()
-        XCTAssert(queue.isEmpty)
+        #expect(queue.isEmpty)
 
         let request1 = MockRequest(connectionType: MockConnection.self)
         queue.queue(request1)
-        XCTAssertEqual(queue.count, 1)
-        XCTAssertFalse(queue.isEmpty)
+        #expect(queue.count == 1)
+        #expect(!queue.isEmpty)
         let popResult = queue.pop(max: 3)
-        XCTAssert(popResult.elementsEqual([request1]))
-        XCTAssert(queue.isEmpty)
-        XCTAssertEqual(queue.count, 0)
+        #expect(popResult.elementsEqual([request1]))
+        #expect(queue.isEmpty)
+        #expect(queue.count == 0)
     }
 
-    func testEnqueueAndPopMultipleRequests() {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testEnqueueAndPopMultipleRequests() {
         var queue = TestQueue()
-        XCTAssert(queue.isEmpty)
+        #expect(queue.isEmpty)
 
         var request1 = MockRequest(connectionType: MockConnection.self)
         queue.queue(request1)
@@ -33,21 +34,22 @@ final class PoolStateMachine_RequestQueueTests: XCTestCase {
         queue.queue(request3)
 
         do {
-            XCTAssertEqual(queue.count, 3)
-            XCTAssertFalse(queue.isEmpty)
+            #expect(queue.count == 3)
+            #expect(!queue.isEmpty)
             let popResult = queue.pop(max: 3)
-            XCTAssert(popResult.elementsEqual([request1, request2, request3]))
-            XCTAssert(queue.isEmpty)
-            XCTAssertEqual(queue.count, 0)
+            #expect(popResult.elementsEqual([request1, request2, request3]))
+            #expect(queue.isEmpty)
+            #expect(queue.count == 0)
         }
-        XCTAssert(isKnownUniquelyReferenced(&request1))
-        XCTAssert(isKnownUniquelyReferenced(&request2))
-        XCTAssert(isKnownUniquelyReferenced(&request3))
+        #expect(isKnownUniquelyReferenced(&request1))
+        #expect(isKnownUniquelyReferenced(&request2))
+        #expect(isKnownUniquelyReferenced(&request3))
     }
 
-    func testEnqueueAndPopOnlyOne() {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testEnqueueAndPopOnlyOne() {
         var queue = TestQueue()
-        XCTAssert(queue.isEmpty)
+        #expect(queue.isEmpty)
 
         var request1 = MockRequest(connectionType: MockConnection.self)
         queue.queue(request1)
@@ -57,24 +59,25 @@ final class PoolStateMachine_RequestQueueTests: XCTestCase {
         queue.queue(request3)
 
         do {
-            XCTAssertEqual(queue.count, 3)
-            XCTAssertFalse(queue.isEmpty)
+            #expect(queue.count == 3)
+            #expect(!queue.isEmpty)
             let popResult = queue.pop(max: 1)
-            XCTAssert(popResult.elementsEqual([request1]))
-            XCTAssertFalse(queue.isEmpty)
-            XCTAssertEqual(queue.count, 2)
+            #expect(popResult.elementsEqual([request1]))
+            #expect(!queue.isEmpty)
+            #expect(queue.count == 2)
 
             let removeAllResult = queue.removeAll()
-            XCTAssert(Set(removeAllResult) == [request2, request3])
+            #expect(Set(removeAllResult) == [request2, request3])
         }
-        XCTAssert(isKnownUniquelyReferenced(&request1))
-        XCTAssert(isKnownUniquelyReferenced(&request2))
-        XCTAssert(isKnownUniquelyReferenced(&request3))
+        #expect(isKnownUniquelyReferenced(&request1))
+        #expect(isKnownUniquelyReferenced(&request2))
+        #expect(isKnownUniquelyReferenced(&request3))
     }
 
-    func testCancellation() {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testCancellation() {
         var queue = TestQueue()
-        XCTAssert(queue.isEmpty)
+        #expect(queue.isEmpty)
 
         var request1 = MockRequest(connectionType: MockConnection.self)
         queue.queue(request1)
@@ -84,34 +87,35 @@ final class PoolStateMachine_RequestQueueTests: XCTestCase {
         queue.queue(request3)
 
         do {
-            XCTAssertEqual(queue.count, 3)
+            #expect(queue.count == 3)
             let returnedRequest2 = queue.remove(request2.id)
-            XCTAssert(returnedRequest2 === request2)
-            XCTAssertEqual(queue.count, 2)
-            XCTAssertFalse(queue.isEmpty)
+            #expect(returnedRequest2 === request2)
+            #expect(queue.count == 2)
+            #expect(!queue.isEmpty)
         }
 
         // still retained by the deque inside the queue
-        XCTAssertEqual(queue.requests.count, 2)
-        XCTAssertEqual(queue.queue.count, 3)
+        #expect(queue.requests.count == 2)
+        #expect(queue.queue.count == 3)
 
         do {
-            XCTAssertEqual(queue.count, 2)
-            XCTAssertFalse(queue.isEmpty)
+            #expect(queue.count == 2)
+            #expect(!queue.isEmpty)
             let popResult = queue.pop(max: 3)
-            XCTAssert(popResult.elementsEqual([request1, request3]))
-            XCTAssert(queue.isEmpty)
-            XCTAssertEqual(queue.count, 0)
+            #expect(popResult.elementsEqual([request1, request3]))
+            #expect(queue.isEmpty)
+            #expect(queue.count == 0)
         }
 
-        XCTAssert(isKnownUniquelyReferenced(&request1))
-        XCTAssert(isKnownUniquelyReferenced(&request2))
-        XCTAssert(isKnownUniquelyReferenced(&request3))
+        #expect(isKnownUniquelyReferenced(&request1))
+        #expect(isKnownUniquelyReferenced(&request2))
+        #expect(isKnownUniquelyReferenced(&request3))
     }
 
-    func testRemoveAllAfterCancellation() {
+    @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
+    @Test func testRemoveAllAfterCancellation() {
         var queue = TestQueue()
-        XCTAssert(queue.isEmpty)
+        #expect(queue.isEmpty)
 
         var request1 = MockRequest(connectionType: MockConnection.self)
         queue.queue(request1)
@@ -121,28 +125,28 @@ final class PoolStateMachine_RequestQueueTests: XCTestCase {
         queue.queue(request3)
 
         do {
-            XCTAssertEqual(queue.count, 3)
+            #expect(queue.count == 3)
             let returnedRequest2 = queue.remove(request2.id)
-            XCTAssert(returnedRequest2 === request2)
-            XCTAssertEqual(queue.count, 2)
-            XCTAssertFalse(queue.isEmpty)
+            #expect(returnedRequest2 === request2)
+            #expect(queue.count == 2)
+            #expect(!queue.isEmpty)
         }
 
         // still retained by the deque inside the queue
-        XCTAssertEqual(queue.requests.count, 2)
-        XCTAssertEqual(queue.queue.count, 3)
+        #expect(queue.requests.count == 2)
+        #expect(queue.queue.count == 3)
 
         do {
-            XCTAssertEqual(queue.count, 2)
-            XCTAssertFalse(queue.isEmpty)
+            #expect(queue.count == 2)
+            #expect(!queue.isEmpty)
             let removeAllResult = queue.removeAll()
-            XCTAssert(Set(removeAllResult) == [request1, request3])
-            XCTAssert(queue.isEmpty)
-            XCTAssertEqual(queue.count, 0)
+            #expect(Set(removeAllResult) == [request1, request3])
+            #expect(queue.isEmpty)
+            #expect(queue.count == 0)
         }
 
-        XCTAssert(isKnownUniquelyReferenced(&request1))
-        XCTAssert(isKnownUniquelyReferenced(&request2))
-        XCTAssert(isKnownUniquelyReferenced(&request3))
+        #expect(isKnownUniquelyReferenced(&request1))
+        #expect(isKnownUniquelyReferenced(&request2))
+        #expect(isKnownUniquelyReferenced(&request3))
     }
 }
