@@ -308,7 +308,6 @@ public final class PostgresClient: Sendable, ServiceLifecycle.Service {
         return try await closure(lease.connection)
     }
 
-    #if compiler(>=6.0)
     /// Lease a connection for the provided `closure`'s lifetime.
     ///
     /// - Parameter closure: A closure that uses the passed `PostgresConnection`. The closure **must not** capture
@@ -316,9 +315,8 @@ public final class PostgresClient: Sendable, ServiceLifecycle.Service {
     /// - Returns: The closure's return value.
     public func withConnection<Result>(
         isolation: isolated (any Actor)? = #isolation,
-        // DO NOT FIX THE WHITESPACE IN THE NEXT LINE UNTIL 5.10 IS UNSUPPORTED
-        // https://github.com/swiftlang/swift/issues/79285
-        _ closure: (PostgresConnection) async throws -> sending Result) async throws -> sending Result {
+        _ closure: (PostgresConnection) async throws -> sending Result
+    ) async throws -> sending Result {
         let lease = try await self.leaseConnection()
 
         defer { lease.release() }
@@ -346,41 +344,13 @@ public final class PostgresClient: Sendable, ServiceLifecycle.Service {
         file: String = #file,
         line: Int = #line,
         isolation: isolated (any Actor)? = #isolation,
-        // DO NOT FIX THE WHITESPACE IN THE NEXT LINE UNTIL 5.10 IS UNSUPPORTED
-        // https://github.com/swiftlang/swift/issues/79285
-        _ closure: (PostgresConnection) async throws -> sending Result) async throws -> sending Result {
-        try await self.withConnection { connection in
-            try await connection.withTransaction(logger: logger, file: file, line: line, closure)
+        _ closure: (PostgresConnection) async throws -> sending Result
+    ) async throws -> sending Result {
+        // for 6.0 to compile we need to explicitly forward the isolation.
+        try await self.withConnection(isolation: isolation) { connection in
+            try await connection.withTransaction(logger: logger, file: file, line: line, isolation: isolation, closure)
         }
     }
-    #else
-
-    /// Lease a connection, which is in an open transaction state, for the provided `closure`'s lifetime.
-    ///
-    /// The function leases a connection from the underlying connection pool and starts a transaction by running a `BEGIN`
-    /// query on the leased connection against the database. It then lends the connection to the user provided closure.
-    /// The user can then modify the database as they wish. If the user provided closure returns successfully, the function
-    /// will attempt to commit the changes by running a `COMMIT` query against the database. If the user provided closure
-    /// throws an error, the function will attempt to rollback the changes made within the closure.
-    ///
-    /// - Parameters:
-    ///   - logger: The `Logger` to log into for the transaction.
-    ///   - file: The file, the transaction was started in. Used for better error reporting.
-    ///   - line: The line, the transaction was started in. Used for better error reporting.
-    ///   - closure: The user provided code to modify the database. Use the provided connection to run queries.
-    ///              The connection must stay in the transaction mode. Otherwise this method will throw!
-    /// - Returns: The closure's return value.
-    public func withTransaction<Result>(
-        logger: Logger,
-        file: String = #file,
-        line: Int = #line,
-        _ closure: (PostgresConnection) async throws -> Result
-    ) async throws -> Result {
-        try await self.withConnection { connection in
-            try await connection.withTransaction(logger: logger, file: file, line: line, closure)
-        }
-    }
-    #endif
 
     /// Run a query on the Postgres server the client is connected to.
     ///
